+++ Backend/Services/Account/AccountService.cs
@@ -1,0 +1,22 @@
+using Backend.Data;
+using Backend.Interfaces.Account;
+using Backend.Models.Entities;
+using Microsoft.EntityFrameworkCore;
+
+namespace Backend.Services.Account;
+
+public class AccountService : IAccountService
+{
+    private readonly ApplicationDbContext _context;
+
+    public AccountService(ApplicationDbContext context)
+    {
+        _context = context;
+    }
+
+    public async Task<User?> LoginAsync(string email, string password)
+    {
+        var existingUser = await _context.Users.FirstOrDefaultAsync(u => u.Email == email && u.PasswordHash == password);
+        return existingUser;
+    }
+}